# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Margen",
    "description": "Asistente financiero personal. Todos los montos viajan como entero de centavos; el servidor no emite decimales en ningún camino monetario.",
    "version": "v1"
  },
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "Movimientos"
        ],
        "operationId": "ListTransactions",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": [
          "Movimientos"
        ],
        "operationId": "GetTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Movimientos"
        ],
        "operationId": "PutTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/rules": {
      "get": {
        "tags": [
          "Reglas"
        ],
        "operationId": "ListRules",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Reglas"
        ],
        "operationId": "CreateRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rules/{id}": {
      "delete": {
        "tags": [
          "Reglas"
        ],
        "operationId": "DeleteRule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRuleRequest": {
        "required": [
          "pattern",
          "matchKind",
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string"
          },
          "matchKind": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 100
          }
        }
      },
      "PutTransactionRequest": {
        "required": [
          "categoryId",
          "status",
          "notes"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "createRule": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  },
...
}